pyproject.toml
[tool.poetry]
name = "docia-app"
version = "0.1.0"
description = ""
authors = ["im"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.348"
huggingface-hub = "^0.19.4"
transformers = "^4.35.2"
Faker = "^20.1.0"
sentence-transformers = "^2.2.2"
bitsandbytes = "^0.41.3"
pgvector = "^0.2.4"
accelerate = "^0.25.0"

fastapi = "^0.104.1"
passlib = { version = "^1.7.4", extras = ["bcrypt"] }
pydantic = "^2.5.1"
uvicorn = "^0.24.0"
python-dotenv = "^1.0.0"
databases = "^0.4.2"
SQLAlchemy-Utils = "^0.41.1"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
uuid = "^1.30"
python-jose = "^3.3.0"
cryptography = "^41.0.5"
annotated-types = "^0.6.0"
python-multipart = "^0.0.6"
pypdf="^4.0.0"

fastapi-pagination="^0.12.14"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
